chore(backup): move service tests into __tests__
Align backup service test layout with the repo co-located __tests__
convention used under src/main/data and other services.
